requirements.txt
gradio==4.27.0
hnswlib==0.8.0
numpy==1.24.1
Pillow==10.3.0
Requests==2.31.0
torch==2.1.2+cu118
tqdm==4.66.1
transformers==4.36.2
